add config file
